record and push files test
