+++ benchmark_record_action/animation.py
@@ -14,21 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-import shutil
-from glob import glob
-from pathlib import Path
 import subprocess
-from benchmark_record_action.config import RESOURCES_DIRECTORY
-import benchmark_record_action.utils.git
-from benchmark_record_action.utils.git import get_current_branch_name
-from benchmark_record_action.utils.webots import get_world_info, expand_world_list, compile_controllers
+import os
 
 
-def _generate_animation_recorder_vrml(duration, output):
+def generate_animation_recorder_vrml(duration, output):
     return (
         f'Robot {{\n'
-        f'  name "supervisor"\n'
+        f'  name "animation_recorder_supervisor"\n'
         f'  controller "animation_recorder"\n'
         f'  controllerArgs [\n'
         f'    "--duration={duration}"\n'
@@ -43,105 +36,38 @@ def _generate_animation_recorder_vrml(duration, output):
         f'}}\n'
     )
 
-
-def _get_animation_directories():
-    directories = []
-    for path in glob('*'):
-        print('path: ', path)
-        if os.path.isdir(path) and os.path.isfile(os.path.join(path, 'index.html')):
-            directories.append(path)
-    return directories
-
-
-def _generate_animation_file(worlds_config):
-    template = None
-
-    # Generate details
-    worlds = [get_world_info(world_config['file']) for world_config in worlds_config]
-
-    # Write to the template
-    with open(os.path.join(RESOURCES_DIRECTORY, 'animation-page.template.html'), 'r') as f:
-        template = f.read()
-    template = template.replace('{ WORLD_LIST_PLACEHOLDER }', str(worlds))
-    with open(os.path.join('/tmp/animation', 'index.html'), 'w') as f:
-        f.write(template)
-
-
-def generate_animation_for_world(world_file, duration, destination_directory='/tmp/animation'):
-    """Generates animation for world given by `world_file`."""
-
-    world_info = get_world_info(world_file)
+def record_animation(world_config, destination_directory):
+    # Get world name and create directory
+    world_name = world_config['file'].split('/')[1]
+    subprocess.check_output(['mkdir', '-p', destination_directory])
 
     # Append `animation_recorder` controller
-    animation_recorder_vrml = _generate_animation_recorder_vrml(
-        duration=duration,
-        output=os.path.join(os.path.abspath('.'), destination_directory, world_info['name'] + '.html')
+    animation_recorder_vrml = generate_animation_recorder_vrml(
+        duration = world_config['duration'],
+        output = os.path.join(os.path.abspath('.'), destination_directory, world_name.replace('.wbt', '.html'))
     )
-    with open(world_file, 'r') as f:
+    with open(world_config['file'], 'r') as f:
         world_content = f.read()
-    with open(world_file, 'w') as f:
+    with open(world_config['file'], 'w') as f:
         f.write(world_content + animation_recorder_vrml)
-    subprocess.check_output(['mkdir', '-p', destination_directory])
 
     # Runs simulation in Webots
     out = subprocess.Popen(
-        ['xvfb-run', 'webots', '--stdout', '--stderr', '--batch', '--mode=fast', '--no-rendering', world_file],
+        ['xvfb-run', 'webots', '--stdout', '--stderr', '--batch', '--mode=fast', '--no-rendering', 'worlds/robot_programming.wbt'],
         stdout=subprocess.PIPE,
         stderr=subprocess.PIPE
     )
+    run_flag = False
     while not out.poll():
         stdoutdata = out.stdout.readline()
         if stdoutdata:
-            print(stdoutdata.decode('utf-8'))
+            if not run_flag: run_flag = True
+            #print(stdoutdata.decode('utf-8'))
         else:
             break
-
     # Removes `animation_recorder` controller
-    with open(world_file, 'w') as f:
+    with open(world_config['file'], 'w') as f:
         f.write(world_content)
 
-
-def generate_animation(animation_config):
-    """Generates animation for all worlds based on the given content."""
-
-    # Add default config
-    if 'worlds' not in animation_config:
-        animation_config['worlds'] = {
-            'file': 'worlds/*.wbt',
-            'duration': 10
-        }
-
-    # Expand world list (handle regex expressions)
-    animation_config['worlds'] = expand_world_list(animation_config['worlds'])
-
-    # Generate animation for each world
-    compile_controllers()
-    for world_config in animation_config['worlds']:
-        generate_animation_for_world(world_config['file'], world_config['duration'])
-
-    # Generates list of animations
-    _generate_animation_file(animation_config['worlds'])
-
-    # Push animation to benchmark-storage branch
-    current_branch_name = get_current_branch_name()
-    benchmark_record_action.utils.git.push_directory_to_branch(
-        '/tmp/animation',
-        destination_directory='AxjD2FU',
-        clean=True
-    )
-
-    # Delete files that are not necessary
-    #animation_directories = _get_animation_directories()
-    """ for path in Path('').glob('*'):
-        path = str(path)
-        print('PATH: ', path)
-        if os.path.isdir(path):
-            shutil.rmtree(path)
-        else:
-            os.remove(path) """
-
-    for path in Path('').glob('*'):
-        path = str(path)
-        print('path: ', path)
-
-    benchmark_record_action.utils.git.push()
+    # Return
+    return run_flag
